# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Киномания",
        "version": "1.0"
    },
    "paths": {
...
        "/api/films/{film}/promo": {
            "post": {
                "tags": [
                    "Promo"
                ],
                "summary": "Set promo status for a film",
                "operationId": "05f80b23f02f16d4ddf6ff7a2a7b1c4b",
                "parameters": [
                    {
                        "name": "film",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "promo"
                                ],
                                "properties": {
                                    "promo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Promo status updated successfully"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        },
        "/api/promo": {
            "get": {
                "tags": [
                    "Promo"
                ],
                "summary": "Get the current promo film",
                "operationId": "ec13d5abafc0b679d83bd55c1fd82518",
                "responses": {
                    "200": {
                        "description": "Current promo film details"
                    },
                    "404": {
                        "description": "Promo film not found"
                    }
                }
            }
        },
...
        "/api/rooms/{room}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a specific room by ID",
                "operationId": "17c2e33477f27ef42f1a6b9942a22832",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a specific room",
                "operationId": "8447e652052d8b92de831d8a566663cd",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a specific room",
                "operationId": "381c2df3391295de921e9c31262d50fd",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Room is deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        },
        "/api/rooms/{room}/join": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Join a room",
                "operationId": "db59d0d88f2c01d281123ac473def630",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User  added to the room",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User  is added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Incorrect password"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User "
                ],
                "summary": "Get the authenticated user details",
                "operationId": "9cd1cee912e0176094a7ec83cc9b432f",
                "responses": {
                    "200": {
                        "description": "User  details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "User "
                ],
                "summary": "Update the authenticated user details",
                "operationId": "fd4f39f7820d3f3cbc259c5b30fed830",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User  updated successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
...
}